requirements.txt
streamlit
pyvis
google-generativeai
python-dotenv
